requirement.txt
convertdate==2.1.0
datefinder==0.6.1
dateparser==0.5.1
Django==1.8.3
django-braces==1.9.0
django-cors-headers==1.1.0
django-filter==0.13.0
django-filters==0.2.1
django-nose==1.4.1
django-oauth-toolkit==0.10.0
django-oauth2-provider==0.2.6.1
django-rest-framework-social-oauth2==1.0.4
djangorestframework==3.3.3
djoser==0.4.3
egenix-mx-base==3.2.9
ephem==3.7.6.0
httplib2==0.10.3
jdatetime==1.8.2
MySQL-python==1.2.5
nltk==3.2.2
nose==1.3.7
numpy
oauth2==1.9.0.post1
oauthlib==1.0.3
PyJWT==1.4.2
python-dateutil==2.6.0
python-openid==2.2.5
python-social-auth==0.2.21
pytz==2016.10
quantities==0.11.1
regex==2016.1.10
requests==2.11.1
requests-oauthlib==0.7.0
ruamel.ordereddict==0.4.9
ruamel.yaml==0.13.14
semantic==1.0.3
shortuuid==0.4.2
six==1.10.0
typing==3.5.3.0
tzlocal==1.3
umalqurra==0.2
vaderSentiment==2.5
